Cover hard-tab insertion, and split the editor test harness
Tab insertion in a tab-indented document had no test. It works — the
hard-tab branch delegates to NSTextView, which inserts a literal tab —
but that was the one indentation path taken on trust, and it is the
branch the width-inference work could most plausibly have disturbed.
Cover it at the start of a line and mid-line.

Adding those pushed EditorIntegrationTests past the type body length
limit, so split it by concern rather than silence the rule: the shared
harness moves to EditorTestCase, highlighting assertions stay in
EditorIntegrationTests, and the editing behaviors — comment toggling,
tab and indent handling, undo, wrap and gutter toggles, font size —
move to EditingIntegrationTests.
