+++ MopedEditor/Tests/MopedEditorTests/EditorIntegrationTests.swift
@@ -22,46 +22,10 @@ import AppKit
 import XCTest
 @testable import MopedEditor
 
-/// Drives a real `MopedTextView` (storage, layout manager, highlighter) the way the
-/// app does, and checks the colors that actually reach the layout manager.
+/// Checks the colors that actually reach the layout manager when a real
+/// `MopedTextView` is driven the way the app drives it.
 @MainActor
-final class EditorIntegrationTests: XCTestCase {
-	/// A windowless text view has no responder chain to inherit an undo manager
-	/// from, so supply one the way the document window does in the app.
-	private final class UndoProvidingDelegate: NSObject, NSTextViewDelegate {
-		let manager = UndoManager()
-
-		func undoManager(for view: NSTextView) -> UndoManager? {
-			manager
-		}
-	}
-
-	private var undoDelegate: UndoProvidingDelegate?
-
-	private func makeEditor(
-		theme: MopedTheme = .defaultLight
-	) -> (scrollView: NSScrollView, textView: MopedTextView) {
-		let editor = MopedTextView.scrollableEditor(theme: theme)
-		editor.scrollView.frame = NSRect(x: 0, y: 0, width: 600, height: 400)
-		let delegate = UndoProvidingDelegate()
-		editor.textView.delegate = delegate
-		undoDelegate = delegate
-		return editor
-	}
-
-	/// Lets the highlighter's coalesced main-queue pass run.
-	private func drainHighlightPasses() {
-		for _ in 0..<10 {
-			RunLoop.current.run(until: Date().addingTimeInterval(0.02))
-		}
-	}
-
-	private func color(at location: Int, in textView: MopedTextView) -> NSColor? {
-		textView.layoutManager?.temporaryAttribute(
-			.foregroundColor, atCharacterIndex: location, effectiveRange: nil
-		) as? NSColor
-	}
-
+final class EditorIntegrationTests: EditorTestCase {
 	func testHighlightsSwiftAfterSettingText() {
 		let (_, textView) = makeEditor()
 		textView.language = "swift"
@@ -201,135 +165,5 @@ final class EditorIntegrationTests: XCTestCase {
 		drainHighlightPasses()
 
 		XCTAssertEqual(textView.undoManager?.canUndo, false, "setPlainText must not create an undo step")
-	}
-
-	func testToggleCommentRoundTrips() {
-		let (_, textView) = makeEditor()
-		textView.language = "swift"
-		textView.lineCommentMarker = "//"
-		textView.setPlainText("let a = 1\nlet b = 2\n")
-		textView.setSelectedRange(NSRange(location: 0, length: (textView.string as NSString).length))
-
-		XCTAssertTrue(textView.toggleLineComment())
-		XCTAssertEqual(textView.string, "// let a = 1\n// let b = 2\n")
-
-		XCTAssertTrue(textView.toggleLineComment())
-		XCTAssertEqual(textView.string, "let a = 1\nlet b = 2\n")
-	}
-
-	func testToggleCommentWithoutMarkerIsRefused() {
-		let (_, textView) = makeEditor()
-		textView.lineCommentMarker = nil
-		textView.setPlainText("plain line\n")
-		textView.setSelectedRange(NSRange(location: 0, length: 0))
-
-		XCTAssertFalse(textView.toggleLineComment())
-		XCTAssertEqual(textView.string, "plain line\n")
-	}
-
-	/// The reported symptom: Tab inserted 2 spaces in a 4-space-indented document
-	/// because the width could never be inferred as 4.
-	func testTabInsertsFourSpacesInFourSpaceDocument() {
-		let (_, textView) = makeEditor()
-		textView.setPlainText("def a():\n    x = 1\n    if x:\n        return x\n")
-		// Caret at the start of the (empty) last line.
-		textView.setSelectedRange(NSRange(location: (textView.string as NSString).length, length: 0))
-
-		textView.insertTab(nil)
-
-		XCTAssertTrue(
-			textView.string.hasSuffix("\n    "),
-			"Expected a 4-space indent, got \(textView.string.suffix(8).debugDescription)"
-		)
-	}
-
-	func testTabInsertsTwoSpacesInTwoSpaceDocument() {
-		let (_, textView) = makeEditor()
-		textView.setPlainText("function a() {\n  if (x) {\n    go();\n  }\n}\n")
-		textView.setSelectedRange(NSRange(location: (textView.string as NSString).length, length: 0))
-
-		textView.insertTab(nil)
-
-		XCTAssertTrue(
-			textView.string.hasSuffix("\n  "),
-			"Expected a 2-space indent, got \(textView.string.suffix(8).debugDescription)"
-		)
-	}
-
-	func testOutdentRemovesFourSpacesInFourSpaceDocument() {
-		let (_, textView) = makeEditor()
-		textView.setPlainText("def a():\n    x = 1\n        y = 2\n")
-		let nsText = textView.string as NSString
-		let lastLine = nsText.range(of: "        y = 2")
-		textView.setSelectedRange(lastLine)
-
-		textView.adjustIndentation(false)
-
-		XCTAssertTrue(
-			textView.string.hasSuffix("\n    y = 2\n"),
-			"Expected one 4-space level removed, got \(textView.string.debugDescription)"
-		)
-	}
-
-	func testIndentAndOutdentSelection() {
-		let (_, textView) = makeEditor()
-		textView.setPlainText("\ta = 1\n\tb = 2\n")
-		textView.setSelectedRange(NSRange(location: 0, length: (textView.string as NSString).length))
-
-		textView.adjustIndentation(true)
-		XCTAssertEqual(textView.string, "\t\ta = 1\n\t\tb = 2\n")
-
-		textView.adjustIndentation(false)
-		XCTAssertEqual(textView.string, "\ta = 1\n\tb = 2\n")
-	}
-
-	func testEditingActionsAreUndoable() {
-		let (_, textView) = makeEditor()
-		textView.lineCommentMarker = "//"
-		textView.setPlainText("let a = 1\n")
-		textView.setSelectedRange(NSRange(location: 0, length: 0))
-
-		XCTAssertTrue(textView.toggleLineComment())
-		XCTAssertEqual(textView.undoManager?.canUndo, true)
-
-		textView.undoManager?.undo()
-		XCTAssertEqual(textView.string, "let a = 1\n", "Toggling a comment must be undoable")
-	}
-
-	func testWrapToggleUpdatesContainerAndScroller() {
-		let (scrollView, textView) = makeEditor()
-		textView.wrapsLines = true
-		XCTAssertEqual(textView.textContainer?.widthTracksTextView, true)
-		XCTAssertFalse(scrollView.hasHorizontalScroller)
-
-		textView.wrapsLines = false
-		XCTAssertEqual(textView.textContainer?.widthTracksTextView, false)
-		XCTAssertTrue(scrollView.hasHorizontalScroller)
-	}
-
-	func testGutterToggle() {
-		let (scrollView, textView) = makeEditor()
-		textView.showsLineNumberGutter = true
-		XCTAssertTrue(scrollView.rulersVisible)
-		XCTAssertTrue(scrollView.verticalRulerView is LineNumberRulerView)
-
-		textView.showsLineNumberGutter = false
-		XCTAssertFalse(scrollView.rulersVisible)
-	}
-
-	func testFontSizeActions() {
-		let (_, textView) = makeEditor()
-		textView.defaultFontSize = 13.0
-		textView.editorFont = NSFont.monospacedSystemFont(ofSize: 13.0, weight: .regular)
-
-		textView.fontSizeIncreaseMenuItemSelected(nil)
-		XCTAssertEqual(textView.editorFont.pointSize, 14.0)
-
-		textView.fontSizeDecreaseMenuItemSelected(nil)
-		XCTAssertEqual(textView.editorFont.pointSize, 13.0)
-
-		textView.fontSizeIncreaseMenuItemSelected(nil)
-		textView.fontSizeResetMenuItemSelected(nil)
-		XCTAssertEqual(textView.editorFont.pointSize, 13.0)
 	}
 }
